requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
requests==2.25.1
Werkzeug==2.0.3